Public API

Contents

Public API

Contents

Index

Index

NeuralQuantumStates.Lattices.Lattice

NeuralQuantumStates.Lattices.Lattice

NeuralQuantumStates.Lattices.Lattice

NeuralQuantumStates.Lattices.Lattice

NeuralQuantumStates.Lattices.Lattice

NeuralQuantumStates.Lattices.LatticeBasis

NeuralQuantumStates.Lattices.LatticeBasis

NeuralQuantumStates.Lattices.LatticeBasis

NeuralQuantumStates.Lattices.LatticeBasis

NeuralQuantumStates.Lattices.LatticeBasis

NeuralQuantumStates.Lattices.LatticeBasis

NeuralQuantumStates.Lattices.build

NeuralQuantumStates.Lattices.build

NeuralQuantumStates.Lattices.build

NeuralQuantumStates.Lattices.build

NeuralQuantumStates.Lattices.build

NeuralQuantumStates.Lattices.Lattice — Type

Lattice{Tᵢ<:Integer,T<:Real,D,O} <: AbstractLattice{Tᵢ,T,D,O}

A D-dimensional NeuralQuantumStates.Lattices.Lattice of shape with given lattice basis and periodic boundary conditions.

Fields

metagraph::MetaGraph{Tᵢ}: A MetaGraphsNext.MetaGraph for the lattice to store its vertices and edges.

shape::SVector{D,Tᵢ}: A vector for the shape of the lattice. It must contain D positive integers.

basis::AbstractLatticeBasis{T,D,O}: A lattice basis for representing the unit cell of the lattice.

periodic::SVector{D,Bool}: A vector for the periodic boundary condition of the lattice in each dimension.

NeuralQuantumStates.Lattices.Lattice — Method

Lattice(
    shape::AbstractVector{Tᵢ},
    basis::AbstractLatticeBasis{T,D,O},
    custom_edges::Tuple{AbstractVector{NTuple{2,L}},AbstractVector{Tᵢ}},
    periodic::SVector{Bool}=fill(false, D)
) where {Tᵢ<:Integer,T<:Real,D,O,L} -> Lattice{Tᵢ,T,D,O}

Build a D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis, periodic boundary conditions and custom_edges.

Arguments

shape::AbstractVector{Tᵢ}: A vector for the shape of the lattice. It must contain D positive integers.

basis::AbstractLatticeBasis{T,D,O}: A lattice basis for representing the unit cell of the lattice.

custom_edges::Tuple{AbstractVector{NTuple{2,L}},AbstractVector{Tᵢ}}: A tuple of two vectors for the custom edges to be added to the lattice. The first vector contains the lattice site labels of the edges to be added in the form of NTuple{2,L} where L is the type of the lattice site labels. The second vector is of positive integers for distingushing the edges to be added. These two vectors must have the same length.

periodic::AbstractVector{Bool}: A vector for the periodic boundary condition of the lattice in each dimension. Defaults to fill(false, D).

Returns

Lattice{Tᵢ,T,D,O}: The built D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis, periodic boundary conditions and custom_edges.

NeuralQuantumStates.Lattices.Lattice — Method

Lattice(
    shape::SVector{D,Tᵢ},
    basis::AbstractLatticeBasis{T,D,O},
    custom_edges::Tuple{AbstractVector{NTuple{2,L}},AbstractVector{Tᵢ}},
    periodic::SVector{D,Bool}=SVector{D,Bool}(fill(false, D))
) where {Tᵢ<:Integer,T<:Real,D,O,L} -> Lattice{Tᵢ,T,D,O}

Build a D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis, periodic boundary conditions and custom_edges.

Arguments

shape::SVector{D,Tᵢ}: A vector for the shape of the lattice. It must contain D positive integers.

basis::AbstractLatticeBasis{T,D,O}: A lattice basis for representing the unit cell of the lattice.

custom_edges::Tuple{AbstractVector{NTuple{2,L}},AbstractVector{Tᵢ}}: A tuple of two vectors for the custom edges to be added to the lattice. The first vector contains the lattice site labels of the edges to be added in the form of NTuple{2,L} where L is the type of the lattice site labels. The second vector is of positive integers for distingushing the edges to be added. These two vectors must have the same length.

periodic::SVector{D,Bool}: A vector for the periodic boundary condition of the lattice in each dimension. Defaults to SVector{D,Bool}(fill(false, D)).

Returns

Lattice{Tᵢ,T,D,O}: The built D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis, periodic boundary conditions and custom_edges.

NeuralQuantumStates.Lattices.Lattice — Method

Lattice(
    shape::AbstractVector{Tᵢ},
    basis::AbstractLatticeBasis{T,D,O},
    periodic::AbstractVector{Bool}=fill(false, D);
    max_order::Tᵢ=1,
    tol_digits::Tᵢ=12,
    dist_tol::T=1.0e-12
) where {Tᵢ<:Integer,T<:Real,D,O} -> Lattice{Tᵢ,T,D,O}

Build a D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis and periodic boundary conditions.

Arguments

shape::AbstractVector{Tᵢ}: A vector for the shape of the lattice. It must contain D positive integers.

basis::AbstractLatticeBasis{T,D,O}: A lattice basis for representing the unit cell of the lattice.

periodic::AbstractVector{Bool}: A vector for the periodic boundary condition of the lattice in each dimension. Defaults to fill(false, D).

Keywords

max_order::Tᵢ: An integer for the maximum order of the edges to be included in the lattice as max_order-nearest neighbors. Defaults to 1, which means only nearest neighbors are included. For example, if it is set to 2, then nearest and next-nearest neighbors are included.

tol_digits::Tᵢ: An integer for the number of digits to round the calculated distances to. Defaults to 12.

dist_tol::T: A positive number for the tolerance of the distance between two lattice sites to be considered as the same site. Defaults to 1.0e-12.

Returns

Lattice{Tᵢ,T,D,O}: The built D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis and periodic boundary conditions.

NeuralQuantumStates.Lattices.Lattice — Method

Lattice(
    shape::SVector{D,Tᵢ},
    basis::AbstractLatticeBasis{T,D,O},
    periodic::SVector{D,Bool}=SVector{D,Bool}(fill(false, D));
    max_order::Tᵢ=1,
    tol_digits::Tᵢ=12,
    dist_tol::T=1.0e-12
) where {Tᵢ<:Integer,T<:Real,D,O} -> Lattice{Tᵢ,T,D,O}

Build a D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis and periodic boundary conditions.

Arguments

shape::SVector{D,Tᵢ}: A vector for the shape of the lattice. It must contain D positive integers.

basis::AbstractLatticeBasis{T,D,O}: A lattice basis for representing the unit cell of the lattice.

periodic::SVector{D,Bool}: A vector for the periodic boundary condition of the lattice in each dimension. Defaults to SVector{D,Bool}(fill(false, D)).

Keywords

max_order::Tᵢ: An integer for the maximum order of the edges to be included in the lattice as max_order-nearest neighbors. Defaults to 1, which means only nearest neighbors are included. For example, if it is set to 2, then nearest and next-nearest neighbors are included.

tol_digits::Tᵢ: An integer for the number of digits to round the calculated distances to. Defaults to 12.

dist_tol::T: A positive number for the tolerance of the distance between two lattice sites to be considered as the same site. Defaults to 1.0e-12.

Returns

Lattice{Tᵢ,T,D,O}: The built D-dimensional NeuralQuantumStates.Lattices.Lattice of shape by using the given lattice basis and periodic boundary conditions.

NeuralQuantumStates.Lattices.LatticeBasis — Type

LatticeBasis{T<:Real,D,O} <: AbstractLatticeBasis{T,D,O}

A lattice basis for representing the unit cell of a D-dimensional lattice with O site offsets.

Fields

vectors::SMatrix{D,D,T}: A D$\times$D square matrix for the primitive vectors defining the unit cell.

site_offsets::SMatrix{D,O,T}: A D$\times$O matrix for the site offsets of the lattice basis in the unit cell.

NeuralQuantumStates.Lattices.LatticeBasis — Method

LatticeBasis(
    vectors::AbstractMatrix{T},
    site_offsets::AbstractMatrix{T}=zeros(T, size(vectors)[1], 1)
) where {T<:Real} -> LatticeBasis{T,size(vectors)[1],size(site_offsets)[2]}

Define a NeuralQuantumStates.Lattices.LatticeBasis for representing the unit cell of a size(vectors)[1]-dimensional lattice with the given primitive vectors and site_offsets.

Arguments

vectors::AbstractMatrix{T}: A square matrix for the primitive vectors defining the unit cell. Each column of the matrix is a primitive vector.

site_offsets::AbstractMatrix{T}: A matrix for the site offsets of the lattice basis in the unit cell. Each column of the matrix is a site offset vector. Defaults to zeros(T, size(vectors)[1], 1).

Returns

LatticeBasis{T,size(vectors)[1],size(site_offsets)[2]}: The defined NeuralQuantumStates.Lattices.LatticeBasis for representing the unit cell of a size(vectors)[1]-dimensional lattice with the given primitive vectors and site_offsets.

NeuralQuantumStates.Lattices.LatticeBasis — Method

LatticeBasis(
    vectors::AbstractMatrix{T},
    site_offset::AbstractVector{T}=zeros(T, size(vectors)[1])
) where {T<:Real} -> LatticeBasis{T,size(vectors)[1],1}

Define a NeuralQuantumStates.Lattices.LatticeBasis for representing the unit cell of a size(vectors)[1]-dimensional lattice with the given primitive vectors and site_offset.

Arguments

vectors::AbstractMatrix{T}: A square matrix for the primitive vectors defining the unit cell. Each column of the matrix is a primitive vector.

site_offset::AbstractVector{T}: A vector for one site offset of the lattice basis in the unit cell. Defaults to zeros(T, size(vectors)[1]).

Returns

LatticeBasis{T,size(vectors)[1],1}: The defined NeuralQuantumStates.Lattices.LatticeBasis for representing the unit cell of a size(vectors)[1]-dimensional lattice with the given primitive vectors and site_offset.

NeuralQuantumStates.Lattices.LatticeBasis — Method

LatticeBasis(
    vectors::AbstractVector{<:AbstractVector{T}},
    site_offsets::AbstractVector{<:AbstractVector{T}}=[zeros(T, length(vectors[1]))]
) where {T<:Real} -> LatticeBasis{T,length(vectors[1]),length(site_offsets)}

Define a NeuralQuantumStates.Lattices.LatticeBasis for representing the unit cell of a length(vectors[1])-dimensional lattice with the given primitive vectors and site_offsets.

Arguments

vectors::AbstractVector{<:AbstractVector{T}}: A vector of primitive vectors defining the unit cell. Each element of the vector should be a primitive vector with the same dimension.

site_offsets::AbstractVector{<:AbstractVector{T}}: A vector of site offsets of the lattice basis in the unit cell. Each element of the vector should be a site offset vector with the same dimension. Defaults to [zeros(T, length(vectors[1]))].

Returns

LatticeBasis{T,length(vectors[1]),length(site_offsets)}: The defined NeuralQuantumStates.Lattices.LatticeBasis for representing the unit cell of a length(vectors[1])-dimensional lattice with the given primitive vectors and site_offsets.

NeuralQuantumStates.Lattices.LatticeBasis — Method

LatticeBasis(
    vector::T, site_offsets::AbstractVector{T}
) where {T<:Real} -> LatticeBasis{T,1,length(site_offsets)}

Define a lattice basis for representing the unit cell of a 1D lattice with length(site_offsets) site offsets.

Arguments

vector::T: A real number for a primitive vector defining the unit cell.

site_offsets::AbstractVector{T}: A vector for the site offsets of the lattice basis in the unit cell.

Returns

LatticeBasis{T,1,length(site_offsets)}: The defined lattice basis of a 1D lattice with length(site_offsets) site offsets.

NeuralQuantumStates.Lattices.LatticeBasis — Method

LatticeBasis(vector::T, site_offset::T=T(0.0)) where {T<:Real} -> LatticeBasis{T,1,1}

Define a NeuralQuantumStates.Lattices.LatticeBasis for representing the unit cell of a 1D lattice with the given primitive vector and site_offset.

Arguments

vector::T: A real number for a primitive vector defining the unit cell.

site_offset::T: A real number for the site offset of the lattice basis in the unit cell. Defaults to T(0.0).

Returns

LatticeBasis{T,1,1}: The defined lattice basis of a 1D lattice with the given primitive vector and site_offset.

NeuralQuantumStates.Lattices.build — Method

build(
    ::Val{:Honeycomb},
    shape::AbstractVector{Tᵢ},
    edge_length::T;
    periodic::Union{Bool,AbstractVector{Bool}}=false,
    tol_digits::Tᵢ=TOL_DIGITS,
    dist_tol::T=DIST_TOL
) where {Tᵢ<:Integer,T<:Real} -> Lattice{Tᵢ,T,2,1}

Build a 2D honeycomb lattice from the given parameters.

Arguments

shape::AbstractVector{Tᵢ}: A vector for the shape of the lattice. It must contain two positive integers.

edge_length::T: A positive number for the edge length of the lattice.

Keywords

periodic::Union{Bool,AbstractVector{Bool}}: A boolean or a vector for the periodic boundary condition of the lattice in each dimension. If it is a boolean, then it is applied to all dimensions. If it is a vector, then it must contain two booleans. Defaults to false.

tol_digits::Tᵢ: An integer for the number of digits to round the calculated distances to. Defaults to 12.

dist_tol::T: A positive number for the tolerance of the distance between two lattice sites to be considered as the same site. Defaults to 1.0e-12.

Returns

Lattice{Tᵢ,T,2,2}: The built 2D honeycomb lattice from the given parameters.

NeuralQuantumStates.Lattices.build — Method

build(
    ::Val{:Hypercube},
    shape::AbstractVector{Tᵢ},
    edge_length::T;
    periodic::Union{Bool,AbstractVector{Bool}}=false,
) where {Tᵢ<:Integer,T<:Real} -> Lattice{Tᵢ,T,length(shape),1}

Build a hypercubic lattice from the given parameters.

Arguments

::Val{:Hypercube}: A value to dispatch to this function.

shape::AbstractVector{Tᵢ}: A vector for the shape of the lattice.

edge_length::T: A positive number for the edge length of the lattice.

Keywords

periodic::Union{Bool,AbstractVector{Bool}}: A boolean or a vector for the periodic boundary condition of the lattice in each dimension. If it is a boolean, then it is applied to all dimensions. If it is a vector, then it must contain length(shape) booleans. Defaults to false.

Returns

Lattice{Tᵢ,T,length(shape),1}: The built hypercubic lattice from the given parameters.

NeuralQuantumStates.Lattices.build — Method

build(
    ::Val{:Kagome},
    shape::AbstractVector{Tᵢ},
    edge_length::T;
    periodic::Union{Bool,AbstractVector{Bool}}=false,
    tol_digits::Tᵢ=TOL_DIGITS,
    dist_tol::T=DIST_TOL
) where {Tᵢ<:Integer,T<:Real} -> Lattice{Tᵢ,T,2,3}

Build a 2D kagome lattice from the given parameters.

Arguments

shape::AbstractVector{Tᵢ}: A vector for the shape of the lattice. It must contain two positive integers.

edge_length::T: A positive number for the edge length of the lattice.

Keywords

periodic::Union{Bool,AbstractVector{Bool}}: A boolean or a vector for the periodic boundary condition of the lattice in each dimension. If it is a boolean, then it is applied to all dimensions. If it is a vector, then it must contain two booleans. Defaults to false.

tol_digits::Tᵢ: An integer for the number of digits to round the calculated distances to. Defaults to 12.

dist_tol::T: A positive number for the tolerance of the distance between two lattice sites to be considered as the same site. Defaults to 1.0e-12.

Returns

Lattice{Tᵢ,T,2,3}: The built 2D kagome lattice from the given parameters.

NeuralQuantumStates.Lattices.build — Method

build(
    ::Val{:Triangular},
    shape::AbstractVector{Tᵢ},
    edge_length::T;
    periodic::Union{Bool,AbstractVector{Bool}}=false,
    tol_digits::Tᵢ=TOL_DIGITS,
    dist_tol::T=DIST_TOL
) where {Tᵢ<:Integer,T<:Real} -> Lattice{Tᵢ,T,2,1}

Build a 2D triangular lattice from the given parameters.

Arguments

::Val{:Triangular}: A value to dispatch to this function.

shape::AbstractVector{Tᵢ}: A vector for the shape of the lattice. It must contain two positive integers.

edge_length::T: A positive number for the edge length of the lattice.

Keywords

periodic::Union{Bool,AbstractVector{Bool}}: A boolean or a vector for the periodic boundary condition of the lattice in each dimension. If it is a boolean, then it is applied to all dimensions. If it is a vector, then it must contain two booleans. Defaults to false.

tol_digits::Tᵢ: An integer for the number of digits to round the calculated distances to. Defaults to 12.

dist_tol::T: A positive number for the tolerance of the distance between two lattice sites to be considered as the same site. Defaults to 1.0e-12.

Returns

Lattice{Tᵢ,T,2,1}: The built 2D triangular lattice from the given parameters.

NeuralQuantumStates.Lattices.build — Method

build(
    ::Val{:Triclinic},
    shape::AbstractVector{Tᵢ},
    edge_lengths::AbstractVector{T},
    angles::AbstractVector{T};
    periodic::Union{Bool,AbstractVector{Bool}}=false,
) where {Tᵢ<:Integer,T<:Real} -> Lattice{Tᵢ,T,3,1}

Build a triclinic lattice from the given parameters.

Arguments

::Val{:Triclinic}: A value to dispatch to this function.

shape::AbstractVector{Tᵢ}: A vector for the shape of the lattice. It must contain three positive integers.

edge_lengths::AbstractVector{T}: A vector for the edge lengths of the lattice. It must contain three positive numbers.

angles::AbstractVector{T}: A vector for the angles in degrees between the edge lengths of the lattice. It must contain three numbers. angles[i] is the angle between edge_lengths[j] and edge_lengths[k] where (i,j,k) is a permutation of (1,2,3).

Keywords

periodic::Union{Bool,AbstractVector{Bool}}: A boolean or a vector for the periodic boundary condition of the lattice in each dimension. If it is a boolean, then it is applied to all dimensions. If it is a vector, then it must contain three booleans. Defaults to false.

Returns

Lattice{Tᵢ,T,3,1}: The built triclinic lattice from the given parameters.